SpreadsheetFormatterSelectorMenu.prepare was SpreadsheetFormatterSelectorMenuList.prepare
