+++ src/main/java/walkingkooka/spreadsheet/server/formatter/SpreadsheetFormatterSelectorMenuContext.java
@@ -20,6 +20,6 @@ package walkingkooka.spreadsheet.server.formatter;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterProvider;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterProviderSamplesContext;
 
-public interface SpreadsheetFormatterSelectorMenuListContext extends SpreadsheetFormatterProvider,
+public interface SpreadsheetFormatterSelectorMenuContext extends SpreadsheetFormatterProvider,
         SpreadsheetFormatterProviderSamplesContext {
 }
