+++ src/test/java/walkingkooka/spreadsheet/server/formatter/SpreadsheetFormatterSelectorMenuTest.java
@@ -17,16 +17,26 @@
 
 package walkingkooka.spreadsheet.server.formatter;
 
+import org.junit.jupiter.api.Test;
+import walkingkooka.collect.list.Lists;
 import walkingkooka.plugin.PluginSelectorMenuLikeTesting;
 import walkingkooka.reflect.JavaVisibility;
+import walkingkooka.spreadsheet.format.SpreadsheetFormatterContext;
+import walkingkooka.spreadsheet.format.SpreadsheetFormatterContextDelegator;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterName;
+import walkingkooka.spreadsheet.format.SpreadsheetFormatterProvider;
+import walkingkooka.spreadsheet.format.SpreadsheetFormatterProviderDelegator;
 import walkingkooka.spreadsheet.format.SpreadsheetFormatterSelector;
+import walkingkooka.spreadsheet.meta.SpreadsheetMetadataTesting;
 import walkingkooka.tree.json.JsonNode;
 import walkingkooka.tree.json.marshall.JsonNodeMarshallingTesting;
 import walkingkooka.tree.json.marshall.JsonNodeUnmarshallContext;
 
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
 public final class SpreadsheetFormatterSelectorMenuTest implements PluginSelectorMenuLikeTesting<SpreadsheetFormatterSelectorMenu, SpreadsheetFormatterSelector, SpreadsheetFormatterName>,
-        JsonNodeMarshallingTesting<SpreadsheetFormatterSelectorMenu> {
+        JsonNodeMarshallingTesting<SpreadsheetFormatterSelectorMenu>,
+        SpreadsheetMetadataTesting {
 
     @Override
     public SpreadsheetFormatterSelectorMenu createPluginSelectorMenu(final String label,
@@ -45,6 +55,108 @@ public final class SpreadsheetFormatterSelectorMenuTest implements PluginSelecto
     @Override
     public SpreadsheetFormatterSelector createDifferentPluginSelector() {
         return SpreadsheetFormatterName.DATE_FORMAT_PATTERN.setText("yyyy/mm");
+    }
+
+    // prepare..........................................................................................................
+
+    @Test
+    public void testPrepareWithContextFails() {
+        assertThrows(
+                NullPointerException.class,
+                () -> SpreadsheetFormatterSelectorMenu.prepare(
+                        null
+                )
+        );
+    }
+
+    @Test
+    public void testPrepare() {
+        this.checkEquals(
+                Lists.of(
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Short",
+                                SpreadsheetFormatterName.DATE_FORMAT_PATTERN.setText("d/m/yy")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Medium",
+                                SpreadsheetFormatterName.DATE_FORMAT_PATTERN.setText("d mmm yyyy")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Long",
+                                SpreadsheetFormatterName.DATE_FORMAT_PATTERN.setText("d mmmm yyyy")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Full",
+                                SpreadsheetFormatterName.DATE_FORMAT_PATTERN.setText("dddd, d mmmm yyyy")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Short",
+                                SpreadsheetFormatterName.DATE_TIME_FORMAT_PATTERN.setText("d/m/yy, h:mm AM/PM")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Medium",
+                                SpreadsheetFormatterName.DATE_TIME_FORMAT_PATTERN.setText("d mmm yyyy, h:mm:ss AM/PM")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Long",
+                                SpreadsheetFormatterName.DATE_TIME_FORMAT_PATTERN.setText("d mmmm yyyy \\a\\t h:mm:ss AM/PM")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Full",
+                                SpreadsheetFormatterName.DATE_TIME_FORMAT_PATTERN.setText("dddd, d mmmm yyyy \\a\\t h:mm:ss AM/PM")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "General",
+                                SpreadsheetFormatterName.GENERAL.setText("")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Number",
+                                SpreadsheetFormatterName.NUMBER_FORMAT_PATTERN.setText("\"#,##0.###\"")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Integer",
+                                SpreadsheetFormatterName.NUMBER_FORMAT_PATTERN.setText("#,##0")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Percent",
+                                SpreadsheetFormatterName.NUMBER_FORMAT_PATTERN.setText("#,##0%")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Currency",
+                                SpreadsheetFormatterName.NUMBER_FORMAT_PATTERN.setText("$#,##0.00")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Default",
+                                SpreadsheetFormatterName.TEXT_FORMAT_PATTERN.setText("@")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Short",
+                                SpreadsheetFormatterName.TIME_FORMAT_PATTERN.setText("h:mm AM/PM")
+                        ),
+                        SpreadsheetFormatterSelectorMenu.with(
+                                "Long",
+                                SpreadsheetFormatterName.TIME_FORMAT_PATTERN.setText("h:mm:ss AM/PM")
+                        )
+                ),
+                SpreadsheetFormatterSelectorMenu.prepare(
+                        new TestSpreadsheetFormatterSelectorMenuContext()
+                ),
+                "prepare"
+        );
+    }
+
+    static class TestSpreadsheetFormatterSelectorMenuContext implements SpreadsheetFormatterSelectorMenuContext,
+            SpreadsheetFormatterProviderDelegator,
+            SpreadsheetFormatterContextDelegator {
+        @Override
+        public SpreadsheetFormatterContext spreadsheetFormatterContext() {
+            return SPREADSHEET_FORMATTER_CONTEXT;
+        }
+
+        @Override
+        public SpreadsheetFormatterProvider spreadsheetFormatterProvider() {
+            return SPREADSHEET_FORMATTER_PROVIDER;
+        }
     }
 
     // json.............................................................................................................
